Revert "Refactor dashboard page"
